Updated the PracticeCard and SmartRevision
Added useQuestionController hook to remove redundancy from PracticeCard
and SmartRevisionCard
