+++ src/features/practice/components/PracticeCard.tsx
@@ -1,231 +1,40 @@
-// 1. Core and external library imports
-// React and router stuff first, pretty standard.
-import { useState, useEffect, useRef, useMemo } from 'react';
-import { useLocation, useNavigate, useParams, useSearchParams } from 'react-router-dom';
-
-// 2. Custom hook imports (application logic)
-// All the complex logic is abstracted into these hooks.
-// Makes this component way cleaner.
-import { useQuestionTimer } from '@/features/questions/hooks/useQuestionTimer';
-import { useQuestionState } from '@/features/questions/hooks/useQuestionState';
-import useQuestionNav from '@/features/questions/hooks/useQuestionNav';
-import useKeyboardShortcuts from '@/shared/hooks/useKeyboardShortcuts';
-import useAnswerFlow from '@/features/questions/hooks/useAnswerFlow';
-
-// 3. Utility and helper function imports
-import { handleBookmark } from '@/features/questions/utils/questionUtils';
-
-// 4. Component imports (UI pieces)
-// Breaking the UI down into smaller, dumber components.
-import ModernLoader from '@/shared/components/ModernLoader';
-import useAuth from '@/shared/hooks/useAuth';
-import useSettings from '@/features/settings/hooks/useSettings';
-import useQuestions from '../hooks/useQuestions';
-import { usePeerBenchmark } from '@/features/questions/hooks/usePeerBenchmark';
-import { toast } from 'sonner';
-import ReportModal from '@/shared/components/ReportModal';
-import QuestionCard from '@/features/questions/components/QuestionCard/QuestionCard';
-import { supabase } from '@/shared/utils/supabaseClient';
+import { useLocation, useParams } from 'react-router-dom';
 import { useGoals } from '@/shared/hooks/useGoals';
-import type { Question } from '@/shared/types/storage';
 
-// This component is the main event for the practice session. It's a "controller" component
-// that pulls together a bunch of hooks and smaller UI components to create the full question view.
+import ModernLoader from '@/shared/components/ModernLoader';
+import QuestionCard from '@/features/questions/components/QuestionCard/QuestionCard';
+import ReportModal from '@/shared/components/ReportModal';
+import useQuestions from '../hooks/useQuestions';
+import { useQuestionController } from '@/features/questions/api/useQuestionController';
+
 const PracticeCard = () => {
-    // Router and Base Navigation Data
-    const navigate = useNavigate();
-    const location = useLocation();
-    const [searchParams] = useSearchParams();
     const { subject, qid } = useParams();
+    const location = useLocation();
     const { subjects } = useGoals();
-    const qs = searchParams.toString();
 
-    const subjectId = subjects.filter((s) => s.slug === subject)[0]?.id;
-    // Fetching questions using useQuestions hook in-case the of shareable-link. For more details visit Issue #6 (https://github.com/Razen04/GateQuest/issues/6)
+    const subjectId = subjects.find((s) => s.slug === subject)?.id;
+
     const { questions: fetchedQuestions, isLoading: isQuestionsLoading } = useQuestions(
         subjectId,
         false,
     );
 
     const passed = location.state?.questions;
-    let questions = Array.isArray(passed) && passed.length ? passed : fetchedQuestions;
+    const questions = Array.isArray(passed) && passed.length ? passed : fetchedQuestions;
 
-    const [currentIndex, setCurrentIndex] = useState<string | number>(qid || 0);
-
-    const currentQuestion = useMemo(() => {
-        if (!questions || questions.length === 0) return null;
-        return (
-            questions.find((q: Question) => String(q.id) === String(currentIndex)) || questions[0]
-        );
-    }, [questions, currentIndex]);
-
-    const safeQuestion =
-        currentQuestion || ({ id: '0', options: [], correct_answer: [], subject: '' } as any);
-
-    // Global contexts
-    const { user, isLogin } = useAuth();
-    const { settings } = useSettings();
-
-    // Question State (User Selection, Numerical Input)
     const {
-        userAnswerIndex,
-        selectedOptionIndices,
-        numericalAnswer,
-        showAnswer,
-        setShowAnswer,
-        result,
-        setResult,
-        resetState: resetQuestionState,
-        handleOptionSelect,
-        handleNumericalInputChange,
-    } = useQuestionState(safeQuestion);
-
-    // Timer Logic
-    const {
-        time: timeTaken,
-        minutes,
-        seconds,
-        isActive: isTimerActive,
-        toggle: toggleTimer,
-        stop: stop,
-    } = useQuestionTimer(settings?.autoTimer, safeQuestion, showAnswer);
-
-    // Answer Flow (Submission Logic)
-    const { handleShowAnswer, handleSubmit } = useAnswerFlow({
-        currentQuestion: safeQuestion,
-        selectedOptionIndices,
-        numericalAnswer,
-        timeTaken,
-        user,
-        isLogin,
-        setShowAnswer,
-        setResult,
-        stop,
-        showAnswer,
+        currentQuestion,
+        isLoading: isControllerLoading,
+        cardProps,
+        modalProps,
+    } = useQuestionController({
+        questions,
+        mode: 'practice',
+        subjectSlug: subject,
+        qid,
     });
 
-    // Navigation Logic
-    const { isFirst, isLast, handleNext, handlePrevious } = useQuestionNav({
-        filteredQuestions: questions,
-        subject,
-        qs,
-        currentIndex, // Pass the ID here
-        setCurrentIndex,
-        resetQuestionState,
-        questionMode: 'practice',
-    });
-
-    // Peer Stats
-    const {
-        benchmarkDetails,
-        loading: statsLoading,
-        message: statsMessage,
-    } = usePeerBenchmark(safeQuestion.id);
-
-    const onExplanationClick = () => {
-        if (safeQuestion.source_url) {
-            window.open(safeQuestion.source_url, '_blank');
-        }
-    };
-
-    // Keyboard Shortcuts
-    useKeyboardShortcuts(
-        {
-            onPrev: handlePrevious,
-            onNext: handleNext,
-            onShowAnswer: handleShowAnswer,
-            onExplain: () => onExplanationClick(),
-        },
-        [safeQuestion],
-    );
-
-    // 7. Side Effects: Audio Handling
-    const correctSoundRef = useRef<HTMLAudioElement | null>(null);
-    const wrongSoundRef = useRef<HTMLAudioElement | null>(null);
-
-    // Initialize Audio
-    useEffect(() => {
-        if (settings?.sound) {
-            correctSoundRef.current = new Audio('/correct.wav');
-            wrongSoundRef.current = new Audio('/wrong.wav');
-        }
-    }, [settings?.sound]);
-
-    // Play Audio on Result
-    useEffect(() => {
-        if (showAnswer && settings?.sound && result !== 'unattempted') {
-            if (result === 'correct') correctSoundRef.current?.play().catch((e) => console.warn(e));
-            else if (result === 'incorrect')
-                wrongSoundRef.current?.play().catch((e) => console.warn(e));
-        }
-    }, [showAnswer, result, settings?.sound]);
-
-    // 8. Event Handlers
-    const [showReportModal, setShowReportModal] = useState(false);
-    const [reportSubmitting, setReportSubmitting] = useState(false);
-
-    const handleReportSubmit = async (reportType: string, reportText: string) => {
-        setReportSubmitting(true);
-        const report = {
-            user_id: user?.id,
-            question_id: currentQuestion.id,
-            report_type: reportType,
-            report_text: reportText,
-        };
-
-        try {
-            const { error } = await supabase.from('question_reports').insert([report]);
-            if (error) {
-                if (error.code === '23505') {
-                    toast.error("Already reported by you, don't spam please");
-                } else {
-                    toast.error('There was an error in submitting the report.');
-                }
-                console.error('Error reporting question:', error);
-                return;
-            }
-            toast.success('Thank you for making the platform great. ❤️');
-            setShowReportModal(false);
-        } catch (err) {
-            console.error(err);
-            toast.error('Unexpected error occurred.');
-        } finally {
-            setReportSubmitting(false);
-        }
-    };
-
-    // Allow sharing question to peers
-    const onShareClick = async () => {
-        if (navigator.share) {
-            try {
-                await navigator.share({
-                    title: 'GATEQuest PYQ question',
-                    text: 'Try out this question:',
-                    url: window.location.href,
-                });
-            } catch (err) {
-                console.error('Share cancelled or failed.', err);
-            }
-        } else {
-            try {
-                await navigator.clipboard.writeText(window.location.href);
-                toast.message('Question link copied successfully.');
-            } catch (err) {
-                console.error(err);
-            }
-        }
-    };
-
-    const onToggleBookmark = () => {
-        handleBookmark(isLogin, safeQuestion.id, safeQuestion.subject);
-    };
-
-    const handleBack = () => {
-        navigate(`/practice/${subject}?${qs}`);
-    };
-
-    // 9. Render Loading
-    if (isQuestionsLoading || !currentQuestion) {
+    if (isQuestionsLoading || isControllerLoading || !currentQuestion) {
         return (
             <div className="flex items-center justify-center h-screen">
                 <ModernLoader />
@@ -235,59 +44,9 @@ const PracticeCard = () => {
 
     return (
         <>
-            <QuestionCard
-                // Data Props
-                question={currentQuestion}
-                totalQuestions={questions.length}
-                questionNumber={
-                    questions.findIndex(
-                        (q: Question) => String(q.id) === String(currentQuestion.id),
-                    ) + 1
-                }
-                // State Props
-                userAnswerIndex={userAnswerIndex}
-                selectedOptionIndices={selectedOptionIndices}
-                numericalAnswer={numericalAnswer}
-                showAnswer={showAnswer}
-                result={result}
-                // Timer Props
-                timer={{
-                    minutes,
-                    seconds,
-                    isActive: isTimerActive,
-                    onToggle: toggleTimer,
-                }}
-                // Stats Props
-                peerStats={{
-                    loading: statsLoading,
-                    message: statsMessage,
-                    data: benchmarkDetails,
-                }}
-                // Event Handlers
-                onOptionSelect={handleOptionSelect}
-                onNumericalChange={handleNumericalInputChange}
-                onShowAnswer={handleShowAnswer}
-                handleSubmit={handleSubmit}
-                onNext={handleNext}
-                onPrev={handlePrevious}
-                onReport={() => setShowReportModal(true)}
-                onShare={onShareClick}
-                onBookmark={onToggleBookmark}
-                onExplanationClick={onExplanationClick}
-                onBack={handleBack}
-                // Navigation Flags
-                isFirst={isFirst}
-                isLast={isLast}
-            />
+            <QuestionCard {...cardProps} />
 
-            {showReportModal && (
-                <ReportModal
-                    show={showReportModal} // Ensure Modal accepts this prop
-                    onClose={() => setShowReportModal(false)}
-                    onSubmit={handleReportSubmit}
-                    reportSubmitting={reportSubmitting}
-                />
-            )}
+            <ReportModal {...modalProps} />
         </>
     );
 };
