+++ src/features/smart-revision/components/SmartRevisionQuestionCard.tsx
@@ -1,23 +1,14 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useLocation, useNavigate, useParams, useSearchParams } from 'react-router-dom';
-import { toast } from 'sonner';
+
+import type { Question } from '@/shared/types/storage';
+
 import QuestionCard from '@/features/questions/components/QuestionCard/QuestionCard';
 import ReportModal from '@/shared/components/ReportModal';
 import ModernLoader from '@/shared/components/ModernLoader';
-import useAnswerFlow from '@/features/questions/hooks/useAnswerFlow';
-import useAuth from '@/shared/hooks/useAuth';
-import useKeyboardShortcuts from '@/shared/hooks/useKeyboardShortcuts';
-import { usePeerBenchmark } from '@/features/questions/hooks/usePeerBenchmark';
-import useQuestionNav from '@/features/questions/hooks/useQuestionNav';
-import { useQuestionState } from '@/features/questions/hooks/useQuestionState';
-import { useQuestionTimer } from '@/features/questions/hooks/useQuestionTimer';
-import useSettings from '@/features/settings/hooks/useSettings';
-import type { Question } from '@/shared/types/storage';
-import { handleBookmark } from '@/features/questions/utils/questionUtils';
-import { supabase } from '@/shared/utils/supabaseClient';
+import { useQuestionController } from '@/features/questions/api/useQuestionController';
 
 const SmartRevisionQuestionCard = () => {
-    // Router and Base Navigation Data
     const navigate = useNavigate();
     const location = useLocation();
     const [searchParams] = useSearchParams();
@@ -38,236 +29,32 @@ const SmartRevisionQuestionCard = () => {
         });
     }, [location.state, qs, navigate, qid, subject]);
 
-    const [currentIndex, setCurrentIndex] = useState<string | number>(qid || 0);
-
-    const currentQuestion = useMemo(() => {
-        if (!questions || questions.length === 0) return null;
-        return (
-            questions.find((q: Question) => String(q.id) === String(currentIndex)) || questions[0]
-        );
-    }, [questions, currentIndex]);
-
-    const safeQuestion =
-        currentQuestion || ({ id: '0', options: [], correct_answer: [], subject: '' } as any);
-
-    // Global contexts
-    const { user, isLogin } = useAuth();
-    const { settings } = useSettings();
-
-    // Timer Logic
     const {
-        time: timeTaken,
-        minutes,
-        seconds,
-        isActive: isTimerActive,
-        toggle: toggleTimer,
-        reset: resetTimer,
-    } = useQuestionTimer(settings?.autoTimer, safeQuestion);
-
-    // Question State (User Selection, Numerical Input)
-    const {
-        userAnswerIndex,
-        selectedOptionIndices,
-        numericalAnswer,
-        showAnswer,
-        setShowAnswer,
-        result,
-        setResult,
-        resetState: resetQuestionState,
-        handleOptionSelect,
-        handleNumericalInputChange,
-    } = useQuestionState(safeQuestion);
-
-    // Answer Flow (Submission Logic)
-    const { handleShowAnswer, handleSubmit } = useAnswerFlow({
-        currentQuestion: safeQuestion,
-        selectedOptionIndices,
-        numericalAnswer,
-        timeTaken,
-        user,
-        isLogin,
-        setShowAnswer,
-        setResult,
-        resetTimer,
-        showAnswer,
-    });
-
-    // Navigation Logic
-    const { isFirst, isLast, handleNext, handlePrevious } = useQuestionNav({
-        filteredQuestions: questions,
-        subject: subject,
-        qs,
-        currentIndex, // Pass the ID here
-        setCurrentIndex,
-        resetQuestionState,
-        questionMode: 'revision',
+        currentQuestion,
+        isLoading: isControllerLoading,
+        cardProps,
+        modalProps,
+    } = useQuestionController({
+        questions,
+        mode: 'revision',
         revisionId: rid,
+        subjectSlug: subject,
+        qid,
     });
 
-    // Peer Stats
-    const {
-        benchmarkDetails,
-        loading: statsLoading,
-        message: statsMessage,
-    } = usePeerBenchmark(safeQuestion.id);
-
-    const onExplanationClick = () => {
-        if (safeQuestion.explanation) {
-            window.open(safeQuestion.explanation, '_blank');
-        }
-    };
-
-    // Keyboard Shortcuts
-    useKeyboardShortcuts(
-        {
-            onPrev: handlePrevious,
-            onNext: handleNext,
-            onShowAnswer: handleShowAnswer,
-            onExplain: () => onExplanationClick(),
-        },
-        [safeQuestion],
-    );
-
-    // 7. Side Effects: Audio Handling
-    const correctSoundRef = useRef<HTMLAudioElement | null>(null);
-    const wrongSoundRef = useRef<HTMLAudioElement | null>(null);
-
-    // Initialize Audio
-    useEffect(() => {
-        if (settings?.sound) {
-            correctSoundRef.current = new Audio('/correct.wav');
-            wrongSoundRef.current = new Audio('/wrong.wav');
-        }
-    }, [settings?.sound]);
-
-    // Play Audio on Result
-    useEffect(() => {
-        if (showAnswer && settings?.sound && result !== 'unattempted') {
-            if (result === 'correct') correctSoundRef.current?.play().catch((e) => console.warn(e));
-            else if (result === 'incorrect')
-                wrongSoundRef.current?.play().catch((e) => console.warn(e));
-        }
-    }, [showAnswer, result, settings?.sound]);
-
-    // 8. Event Handlers
-    const [showReportModal, setShowReportModal] = useState(false);
-
-    const handleReportSubmit = async (reportType: string, reportText: string) => {
-        const report = {
-            user_id: user?.id,
-            question_id: currentQuestion?.id,
-            report_type: reportType,
-            report_text: reportText,
-        };
-
-        const { error } = await supabase.from('question_reports').insert([report]);
-
-        if (error) {
-            if (error.code === '23505') {
-                toast.error("Already reported by you, don't spam please");
-            } else {
-                toast.error('There was an error in submitting the report.');
-            }
-            console.error('Error reporting question:', error);
-        } else {
-            toast.success('Thank you for making the platform great. ❤️');
-        }
-
-        setShowReportModal(false);
-    };
-
-    // Allow sharing question to peers
-    const onShareClick = async () => {
-        if (navigator.share) {
-            try {
-                await navigator.share({
-                    title: 'GATEQuest PYQ question',
-                    text: 'Try out this question:',
-                    url: window.location.href,
-                });
-            } catch (err) {
-                console.error('Share cancelled or failed.', err);
-            }
-        } else {
-            try {
-                await navigator.clipboard.writeText(window.location.href);
-                toast.message('Question link copied successfully.');
-            } catch (err) {
-                console.error(err);
-            }
-        }
-    };
-
-    const onToggleBookmark = () => {
-        handleBookmark(isLogin, safeQuestion.id, safeQuestion.subject);
-    };
-
-    const handleBack = () => {
-        navigate(`/revision/${rid}?${qs}`);
-    };
-
-    // 9. Render Loading or Dumb Component
-    if (!questions || !currentQuestion) {
+    if (!questions.length || isControllerLoading || !currentQuestion) {
         return (
             <div className="flex items-center justify-center h-screen">
                 <ModernLoader />
             </div>
         );
     }
+
     return (
         <div>
-            <QuestionCard
-                // Data Props
-                question={currentQuestion}
-                totalQuestions={questions.length}
-                questionNumber={
-                    questions.findIndex(
-                        (q: Question) => String(q.id) === String(currentQuestion.id),
-                    ) + 1
-                }
-                // State Props
-                userAnswerIndex={userAnswerIndex}
-                selectedOptionIndices={selectedOptionIndices}
-                numericalAnswer={numericalAnswer}
-                showAnswer={showAnswer}
-                result={result}
-                // Timer Props
-                timer={{
-                    minutes,
-                    seconds,
-                    isActive: isTimerActive,
-                    onToggle: toggleTimer,
-                }}
-                // Stats Props
-                peerStats={{
-                    loading: statsLoading,
-                    message: statsMessage,
-                    data: benchmarkDetails,
-                }}
-                // Event Handlers
-                onOptionSelect={handleOptionSelect}
-                onNumericalChange={handleNumericalInputChange}
-                onShowAnswer={handleShowAnswer}
-                handleSubmit={handleSubmit}
-                onNext={handleNext}
-                onPrev={handlePrevious}
-                onReport={() => setShowReportModal(true)}
-                onShare={onShareClick}
-                onBookmark={onToggleBookmark}
-                onExplanationClick={onExplanationClick}
-                onBack={handleBack}
-                // Navigation Flags
-                isFirst={isFirst}
-                isLast={isLast}
-            />
+            <QuestionCard {...cardProps} />
 
-            {showReportModal && (
-                <ReportModal
-                    show={showReportModal} // Ensure Modal accepts this prop
-                    onClose={() => setShowReportModal(false)}
-                    onSubmit={handleReportSubmit}
-                />
-            )}
+            <ReportModal {...modalProps} />
         </div>
     );
 };
